4_user_roles add pundit, seeds, rpsec controller sign-in

+++ app/controllers/wikis_controller.rb
@@ -1,42 +1,44 @@
 class WikisController < ApplicationController
+  before_action :authenticate_user!, only: [:new, :create, :edit, :update, :destroy]
   before_action :set_wiki, only: [:show, :edit, :update, :destroy]
-#  before_action :authenticate_user!, only: [:new, :create, :edit, :update]
 
   def index
-    @wikis = Wiki.all
+    #@wikis = Wiki.all
+    #authorize @wikis
+    @wikis = policy_scope(Wiki)
   end
 
   def show
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
-# no flash messges
+
   def create
-    #byebug
     @wiki = Wiki.new(wiki_params)
-    #byebug
     @wiki.user = current_user
+    authorize @wiki
     #byebug
     if @wiki.save
-      flash[:notice] = "Your wiki was saved successfully!"
+      flash[:success] = "Your wiki was saved successfully!"
       redirect_to(@wiki)
     else
       flash[:error] = "Your wiki was not saved. Please try again."
       render :new
-
-
-    #  byebug
     end
   end
 
   def edit
+    authorize @wiki
   end
 
   def update
+    authorize @wiki
     if @wiki.update_attributes(wiki_params)
-      flash[:notice] = "Your wiki was updated successfully!"
+      flash[:success] = "Your wiki was updated successfully!"
       redirect_to(@wiki)
     else
       flash[:error] = "Your wiki was not updated. Please try again."
@@ -45,9 +47,10 @@ class WikisController < ApplicationController
   end
 
   def destroy
+    authorize @wiki
     if @wiki.destroy
+      flash[:notice] = "Your wiki was deleted."
       redirect_to wikis_url
-      #byebug
     end
   end
